create redis config and import to index
